🗄️ Move posts container componentes to context components
