+++ src/app/components/PostsList/styles.module.scss
