--- src/app/components/ui/PostsViewSelector/styles.module.scss
@@ -1,6 +1,0 @@
-@import "@/styles/_spacing.scss";
-
-.postsViewSelector {
-  display: flex;
-  gap: calc($unit * 2);
-}
